\name{DEGRE-package}
\alias{DEGRE-package}
\alias{DEGRE}
\docType{package}
\title{
\packageTitle{DEGRE}
}
\description{
\packageDescription{DEGRE}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{DEGRE}
\packageIndices{DEGRE}
The DEGRE package has the function DEGRE with the arguments: count_matrix, num_reps, p_value_adjustment = "BH", formula, design_matrix.
The user needs to have the count matrix of the genes and the design matrix as data frames to enter in the respective arguments.
}
\author{
\packageAuthor{DEGRE}

Maintainer: \packageMaintainer{DEGRE}
}
\references{
Machado et al, 2022
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
%~~ the R documentation directory ~~
\keyword{ package }
\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%~~ simple examples of the most important functions ~~
}
